+++ Pandora/Paradigm/Structure/Specific.hs
@@ -1,6 +1,5 @@
-module Pandora.Paradigm.Structure (module Exports) where
+module Pandora.Paradigm.Structure.Specific (module Exports) where
 
-import Pandora.Paradigm.Structure.Nonempty as Exports
 import Pandora.Paradigm.Structure.Specific.Binary as Exports
 import Pandora.Paradigm.Structure.Specific.Graph as Exports
 import Pandora.Paradigm.Structure.Specific.Stack as Exports